# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SolveX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new crypto order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/confirm:
    post:
      operationId: confirmOrder
      tags: [orders]
      summary: Confirm payment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmOrderInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  # ── Vault ─────────────────────────────────────────────────────────────
...
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: List problems with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Problems list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
    post:
      operationId: createProblem
      tags: [problems]
      summary: Post a new problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemInput"
      responses:
        "201":
          description: Problem created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /problems/{id}:
    get:
      operationId: getProblem
      tags: [problems]
      summary: Get problem by ID with solutions and escrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problem detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /problems/{id}/status:
    patch:
      operationId: updateProblemStatus
      tags: [problems]
      summary: Update problem status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  # ── Solutions ─────────────────────────────────────────────────────────
...
  /owner/audit-log:
    get:
      operationId: getAuditLog
      tags: [owner]
      summary: Get audit log (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
...
  /owner/withdraw:
    post:
      operationId: initiateOwnerWithdrawal
      tags: [owner]
      summary: Initiate owner withdrawal (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerWithdrawInput"
      responses:
        "200":
          description: Withdrawal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerWithdrawResult"
...
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    OrderInput:
      type: object
      required: [productId, paymentMethod]
      properties:
        productId:
          type: string
        paymentMethod:
          type: string
          enum: [eth, usdc, btc]
    OrderResult:
      type: object
      required: [orderId, walletAddress, amount, paymentMethod]
      properties:
        orderId:
          type: string
        walletAddress:
          type: string
        amount:
          type: string
        paymentMethod:
          type: string
        holdUntil:
          type: string
    Order:
      type: object
      required: [id, userId, productId, status, paymentMethod, amount, walletAddress]
      properties:
        id:
          type: string
        userId:
          type: integer
        productId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, delivered, failed]
        paymentMethod:
          type: string
        amount:
          type: string
        walletAddress:
          type: string
        transactionHash:
          type: ["string", "null"]
        createdAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
    ConfirmOrderInput:
      type: object
      required: [transactionHash]
      properties:
        transactionHash:
          type: string
...
    Problem:
      type: object
      required: [id, title, description, category, status, paymentOffer]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        source:
          type: string
        sourceUrl:
          type: ["string", "null"]
        paymentOffer:
          type: string
        currency:
          type: string
        deadline:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        tags:
          type: ["string", "null"]
        viewCount:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ProblemInput:
      type: object
      required: [title, description, category, paymentOffer]
      properties:
        title:
          type: string
          minLength: 10
        description:
          type: string
          minLength: 20
        category:
          type: string
        paymentOffer:
          type: number
          minimum: 1
        deadline:
          type: string
        tags:
          type: array
          items:
            type: string
    ProblemStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [closed, refunded]
...
    ProblemDetail:
      type: object
      required: [problem]
      properties:
        problem:
          $ref: "#/components/schemas/Problem"
        solutions:
          type: array
          items:
            $ref: "#/components/schemas/Solution"
        escrow:
          $ref: "#/components/schemas/Escrow"
    Solution:
      type: object
      required: [id, problemId, solverId, content, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        solverId:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [pending, verifying, approved, rejected]
        verificationScore:
          type: ["string", "null"]
        verificationNotes:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Escrow:
      type: object
      required: [id, problemId, clientId, amount, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        clientId:
          type: integer
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, held, released, refunded, failed]
        stripePaymentIntentId:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
        refundedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AuditLogEntry:
      type: object
      required: [id, eventType, status, createdAt]
      properties:
        id:
          type: string
        eventType:
          type: string
        userId:
          type: ["integer", "null"]
        orderId:
          type: ["string", "null"]
        details:
          type: string
        status:
          type: string
        createdAt:
          type: string
...
    OwnerWithdrawInput:
      type: object
      required: [entryIds, withdrawalAddress]
      properties:
        entryIds:
          type: array
          items:
            type: string
        withdrawalAddress:
          type: string
    OwnerWithdrawResult:
      type: object
      required: [success, totalAmount, entryCount]
      properties:
        success:
          type: boolean
        totalAmount:
          type: string
        entryCount:
          type: integer